+++ src/components/Map.tsx
@@ -1,24 +1,35 @@
 import React from 'react'
-import { Map as LeafletMap } from '@datapunt/react-maps'
+import { Map as ReactMap } from '@datapunt/react-maps'
+import styled, { css } from 'styled-components'
 import { LeafletEventHandlerFnMap, MapOptions } from 'leaflet'
 import { DEFAULT_AMSTERDAM_MAPS_OPTIONS } from '../constants'
 
-export interface Props {
+type StyleProps = {
+  fullScreen?: boolean
+}
+
+export type Props = {
   events?: LeafletEventHandlerFnMap
   options?: MapOptions
   setInstance?: (instance: L.Map) => void
-}
+} & StyleProps
+
+const StyledMap = styled((
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  { fullScreen, ...otherProps }, // make sure we don't pass the StyleProps to the ReactMap component
+) => <ReactMap {...otherProps} />)<StyleProps>`
+  ${({ fullScreen }) =>
+    fullScreen &&
+    css`
+      width: 100%;
+      height: 100vh;
+    `}
+`
 
 const Map: React.FC<Props> = ({
   children,
   options = DEFAULT_AMSTERDAM_MAPS_OPTIONS,
-  events,
-  setInstance,
   ...otherProps
-}) => (
-  <LeafletMap {...{ options, setInstance, events, ...otherProps }}>
-    {children}
-  </LeafletMap>
-)
+}) => <StyledMap {...{ options, ...otherProps }}>{children}</StyledMap>
 
 export default Map