+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Prefix the change with one of these keywords:
 - _Fixed_: for any bug fixes.
 - _Security_: in case of vulnerabilities.
 
+## UNRELEASED
+
+- Added: RDGeoJSON
+
 ## [0.2.0]
 
 - Added: Draw tool